refactor(slacksvc): move chat status-line seam to notifier_chat.go (PRD #1026 M1)
